+++ R/sensor_tracking.R
@@ -7,15 +7,6 @@
 #' it is recommended to use large chunks and large buffers (e.g. a swath width). The point cloud must
 #' not be normalized.
 #'
-#' When multiple returns from a single pulse are detected, the sensor computes their positions as being
-#' in the center of the footprint and thus all aligned. Because of that behavior, a line
-#' drawn between and beyond those returns must cross the sensor. Thus, several consecutive pulses
-#' emitted in a tight interval (e.g. 0.5 seconds) can be used to approximate an intersection
-#' point in the sky that corresponds to the sensor position given that the sensor carrier hasn't
-#' moved much during this interval. A weighted least squares method gives an approximation of the
-#' intersection by minimizing the squared sum of the distances between the intersection point and all
-#' the lines.
-#'
 #' @section Test of data integrity:
 #' In theory, sensor tracking is a simple problem to solve as long as each pulse is properly
 #' identified from a well-populated dataset. In practice, many problems may arise from datasets that are populated
@@ -36,11 +27,8 @@
 #' @template section-supported-options-sensor_tracking
 #'
 #' @template param-las
-#' @param interval numeric. Interval used to bin the gps times and group the pulses to compute
-#' a position at a given timepoint t.
-#' @param pmin integer. Minimum number of pulses needed to estimate a sensor position.
-#' For a given interval, the sensor position is not computed if the number of pulses is lower than
-#' \code{pmin}.
+#' @param algorithm function. An algorithm to compute sensor tracking. \code{lidR} implements
+#' \link{Roussel2020} and  \link{Gatziolis2019} (see respective documentation and examples).
 #' @param extra_check boolean. Datasets are rarely perfectly populated, leading to unexpected errors.
 #' Time-consuming checks of data integrity are performed. These checks can be skipped as they account
 #' for an important proportion of the computation time. See also section 'Tests of data integrity'.
@@ -48,9 +36,11 @@
 #' multiple returns. It is more computationally demanding but not necessarily more accurate. This keeps
 #' only one pulse every x seconds. Set to 0 to use all multiple returns. Use 0 if the file has already
 #' been read with \code{filter = "-thin_pulses_with_time 0.001"}.
+#' @param multi_pulse logical. TRUE only for systems with multiple pulses. Pulse ID must be recorded
+#' in the UserData attribute.
 #' @return A SpatialPointsDataFrame with the Z elevation stored in the table of attributes. Information
-#' about the time interval and the number of pulses used to find the points is also in the table of
-#' attributes.
+#' about the time interval and the score of the positionning (according the the method used) are also
+#' in the table of attributes.
 #'
 #' @author Jean-Francois Bourdon & Jean-Romain Roussel
 #' @export
@@ -63,7 +53,7 @@
 #' # pmin = 15 because it is an extremely tiny file
 #' # strongly decimated to reduce its size. There are
 #' # actually few multiple returns
-#' flightlines <- track_sensor(las, pmin = 15)
+#' flightlines <- track_sensor(las, Roussel2020(pmin = 15))
 #'
 #' plot(las@header)
 #' plot(flightlines, add = TRUE)
@@ -75,7 +65,7 @@
 #' las <- readLAS(LASfile,
 #'                select = "xyzrntp",
 #'                filter = "-drop_single -thin_pulses_with_time 0.001")
-#' flightlines <- track_sensor(las)
+#' flightlines <- track_sensor(las,  Roussel2020(pmin = 15))
 #'
 #' x <- plot(las)
 #' add_flightlines3d(x, flightlines, radius = 10)
@@ -84,17 +74,17 @@
 #' # With a LAScatalog "-drop_single" and "-thin_pulses_with_time"
 #' # are used by default
 #' ctg = readLAScatalog("folder/")
-#' flightlines <- track_sensor(ctg)
+#' flightlines <- track_sensor(ctg,  Roussel2020(pmin = 15))
 #' plot(flightlines)
 #' }
 #' @family range
-track_sensor <- function(las, interval = 0.5, pmin = 50, extra_check = TRUE, thin_pulse_with_time = 0.001)
+track_sensor <- function(las, algorithm, extra_check = TRUE, thin_pulse_with_time = 0.001, multi_pulse = FALSE)
 {
   UseMethod("track_sensor", las)
 }
 
 #' @export
-track_sensor.LAS <- function(las, interval = 0.5, pmin = 50, extra_check = TRUE, thin_pulse_with_time = 0.001)
+track_sensor.LAS <- function(las, algorithm, extra_check = TRUE, thin_pulse_with_time = 0.001, multi_pulse = FALSE)
 {
   if (!"PointSourceID" %in% names(las@data))     stop("No 'PointSourceID' attribute found", call. = FALSE)
   if (!"gpstime" %in% names(las@data))           stop("No 'gpstime' attribute found", call. = FALSE)
@@ -105,12 +95,21 @@ track_sensor.LAS <- function(las, interval = 0.5, pmin = 50, extra_check = TRUE,
   if (!any(las@data[["NumberOfReturns"]] != 0))  stop("'NumberOfReturns' attribute is not populated.", call. = FALSE)
   if (!any(las@data[["PointSourceID"]] != 0))    stop("'PointSourceID' attribute is not populated.", call. = FALSE)
 
-  assert_is_a_number(interval)
-  assert_is_a_number(pmin)
+  if (multi_pulse)
+  {
+    if (!"UserData" %in% names(las@data))
+      stop("No 'UserData' attribute found", call. = FALSE)
+  }
+
+  assert_is_algorithm(algorithm)
+  assert_is_algorithm_trk(algorithm)
   assert_is_a_bool(extra_check)
   assert_is_a_number(thin_pulse_with_time)
+  assert_is_a_bool(multi_pulse)
 
-  . <- X <- Y <- Z <- ReturnNumber <- NumberOfReturns <- PointSourceID <- pulseID <- gpstime <-  NULL
+  lidR.context = "track_sensor"
+
+  . <- X <- Y <- Z <- ReturnNumber <- NumberOfReturns <- PointSourceID <- pulseID <- gpstime <- UserData <-  NULL
 
   data <- las@data
 
@@ -126,7 +125,11 @@ track_sensor.LAS <- function(las, interval = 0.5, pmin = 50, extra_check = TRUE,
   }
 
   # Compute an ID for each pulse
-  data.table::setorder(data, gpstime, ReturnNumber)
+  if (multi_pulse == FALSE)
+    data.table::setorder(data, gpstime, ReturnNumber)
+  else
+    data.table::setorder(data, UserData, gpstime, ReturnNumber) ## set record order
+
   data$pulseID <- .lagisdiff(data[["gpstime"]])
 
   # Count the point per pulse. We kept only first and last so it should be always 2
@@ -137,7 +140,7 @@ track_sensor.LAS <- function(las, interval = 0.5, pmin = 50, extra_check = TRUE,
   if (more_than_2 > 0)
     stop(glue::glue("After keeping only first and last returns of multiple returns pulses, {more_than_2} pulses still have more than 2 points. This dataset is corrupted and gpstime is likely to be invalid."), call. = FALSE)
 
-  # Some edge point might no be paired. Removed them.
+  # Some edge point might not be paired. Removed them.
   ii    <- which(count == 1L)
   data  <- data[!pulseID %in% ii]
 
@@ -160,11 +163,7 @@ track_sensor.LAS <- function(las, interval = 0.5, pmin = 50, extra_check = TRUE,
     data  <- data[!pulseID %in% ii]
   }
 
-  # Generate the bins
-  bins <- round_any(data$gpstime, interval)
-
-  # Find the position P of the sensor in each interval
-  P  <- data[, if (.N > 2*pmin) sensor_positions(X,Y,Z, ReturnNumber), by = .(gpstime = bins, PointSourceID = PointSourceID)]
+  P <- algorithm(data)
 
   # If no position found return an empty SpatialPointsDataFrame
   if (nrow(P) == 0)
@@ -172,7 +171,7 @@ track_sensor.LAS <- function(las, interval = 0.5, pmin = 50, extra_check = TRUE,
     i <- integer(1)
     n <- numeric(1)
     coord <- matrix(i, ncol = 2)
-    data  <- data.frame(Z = i, gpstime = i, PointSourceID = n, npulses = n)
+    data  <- data.frame(Z = i, gpstime = i, PointSourceID = n, SCORE = n)
     zero  <- sp::SpatialPointsDataFrame(coord, data, proj4string = las@proj4string)
     zero  <- zero[-1,]
     return(zero)
@@ -189,20 +188,24 @@ track_sensor.LAS <- function(las, interval = 0.5, pmin = 50, extra_check = TRUE,
 }
 
 #' @export
-track_sensor.LAScluster <- function(las, interval = 0.5, pmin = 50, extra_check = TRUE, thin_pulse_with_time = 0.001)
+track_sensor.LAScluster <- function(las, algorithm, extra_check = TRUE, thin_pulse_with_time = 0.001, multi_pulse = FALSE)
 {
   x <- readLAS(las)
   if (is.empty(x)) return(NULL)
-  pos  <- track_sensor(x, interval, pmin, extra_check)
+  pos  <- track_sensor(x, algorithm, extra_check, thin_pulse_with_time, multi_pulse)
   return(pos)
 }
 
 #' @export
-track_sensor.LAScatalog <- function(las, interval = 0.5, pmin = 50, extra_check = TRUE, thin_pulse_with_time = 0.001)
+track_sensor.LAScatalog <- function(las, algorithm, extra_check = TRUE, thin_pulse_with_time = 0.001, multi_pulse = FALSE)
 {
   assert_is_a_number(thin_pulse_with_time)
 
-  opt_select(las) <- "xyzrntp"
+  if (multi_pulse == FALSE)
+    opt_select(las) <- "xyzrntpa"
+  else
+    opt_select(las) <- "xyzrntpau"
+
   opt_filter(las) <- paste("-drop_single -thin_pulses_with_time", format(thin_pulse_with_time, scientific =  FALSE), opt_filter(las))
 
   if (opt_output_files(las) != "")
@@ -212,71 +215,16 @@ track_sensor.LAScatalog <- function(las, interval = 0.5, pmin = 50, extra_check 
   }
 
   options <- list(need_buffer = TRUE, drop_null = TRUE, need_output_file = FALSE)
-  output  <- catalog_apply(las, track_sensor, interval = interval, pmin = pmin, extra_check = extra_check, thin_pulse_with_time = 0, .options = options)
+  output  <- catalog_apply(las, track_sensor, algorithm = algorithm, extra_check = extra_check, thin_pulse_with_time = 0, multi_pulse = multi_pulse, .options = options)
   output  <- do.call(rbind, output)
   output@proj4string <- las@proj4string
 
   # Post-processing to remove duplicated positions selecting the one computed with the more pulses
-  npulses <- gpstime <- NULL
+  SCORE <- gpstime <- NULL
   data <- as.data.frame(output)
   data.table::setDT(data)
-  i <- data[, .I[which.max(npulses)], by = gpstime]$V1
+  i <- data[, .I[which.max(SCORE)], by = gpstime]$V1
+  data.table::setDF(data)
 
   return(output[i,])
-}
-
-sensor_positions <- function(x, y, z, rn)
-{
-  first <- rn == 1L
-  last  <- rn > 1L
-
-  # Adapted from MATLAB (Anders Eikenes, 2012)
-  # http://www.mathworks.com/matlabcentral/fileexchange/37192-intersection-point-of-lines-in-3d-space
-
-  # Matrix n x 3 containing XYZ coordinates for each first/last return
-  A <- matrix(c(x[first], y[first], z[first]), ncol = 3L)
-  B <- matrix(c(x[last], y[last], z[last]), ncol = 3L)
-
-  # This should never happen but it might happen in wrongly populated datasets
-  if (nrow(A) != nrow(B))
-    return(list(X = NA_real_, Y = NA_real_, Z = NA_real_, npulses = NA_integer_))
-
-  # Lines described as vectors V
-  V <- B - A
-
-  # Weights lines by distances
-  D <- matrix(sqrt(.rowSums(V^2, nrow(V), ncol(V))))
-  W <- D / sum(D)
-
-  # Normalized vectors (|V| = 1)
-  ones <- matrix(1, ncol = 3)
-  N  <- V / (D %*% ones)
-  Nx <- N[, 1L, drop = FALSE]
-  Ny <- N[, 2L, drop = FALSE]
-  Nz <- N[, 3L, drop = FALSE]
-
-  Wxx <- W * (Nx^2 - 1)
-  Wyy <- W * (Ny^2 - 1)
-  Wzz <- W * (Nz^2 - 1)
-  Wxy <- W * Nx * Ny
-  Wxz <- W * Nx * Nz
-  Wyz <- W * Ny * Nz
-
-  Sxx <- sum(Wxx)
-  Syy <- sum(Wyy)
-  Szz <- sum(Wzz)
-  Sxy <- sum(Wxy)
-  Sxz <- sum(Wxz)
-  Syz <- sum(Wyz)
-  S   <- matrix(c(Sxx, Sxy, Sxz, Sxy, Syy, Syz, Sxz, Syz, Szz), nrow = 3L)
-
-  Cx <- sum(A[,1, drop = FALSE] * Wxx + A[,2, drop = FALSE] * Wxy + A[,3, drop = FALSE] * Wxz)
-  Cy <- sum(A[,1, drop = FALSE] * Wxy + A[,2, drop = FALSE] * Wyy + A[,3, drop = FALSE] * Wyz)
-  Cz <- sum(A[,1, drop = FALSE] * Wxz + A[,2, drop = FALSE] * Wyz + A[,3, drop = FALSE] * Wzz)
-  C  <- matrix(c(Cx,Cy,Cz))
-
-  M  <- t(solve(S,C))
-  M  <- round(M, 3)
-
-  return(list(X = M[1], Y = M[2], Z = M[3], npulses = nrow(A)))
 }